// src/models/VentasTiempo.ts
export class VentasTiempo {
    id: any;
    monto_eventual_sin_descuento: any;
    monto_eventual: any;
    costo_eventual: any;
    ganancia: any;
    fecha: any;
    hora: any;
    cajero: any;

    constructor() {
        this.id = "";
        this.monto_eventual_sin_descuento = "";
        this.monto_eventual = "";
        this.costo_eventual = "";
        this.ganancia = "";
        this.fecha = "";
        this.hora = "";
        this.cajero = "";
    }
}