// src/models/MostrarCarrito.ts
export class MostrarCarrito {
    producto: any;
    descripcion: any;
    cantidad: any;
    monto_eventual_sin_descuento: any;
    descuento: any;
    monto_eventual: any;
    id_carrito: any;
    id_producto: any;
    limite: any;

    constructor() {
        this.producto = "";
        this.descripcion = "";
        this.cantidad = "";
        this.monto_eventual_sin_descuento = "";
        this.descuento = "";
        this.monto_eventual = "";
        this.id_carrito = "";
        this.id_producto = "";
        this.limite = "";
    }
}